cleanup

+++ lib/palettes.cpp
@@ -60,8 +60,6 @@ PaletteState::PaletteState()
 {
    // Initialize the palette textures (will create new texture ids)
    InitTextures();
-   // Generate the textures
-   // GenerateTextures();
 }
 
 void PaletteState::InitTextures()
@@ -90,7 +88,7 @@ void PaletteState::InitTextures()
 
 void PaletteState::GenerateTextures()
 {
-   if (textures.size() != Palettes->NumPalettes())
+   if ((int)(textures.size()) != Palettes->NumPalettes())
    {
       cout << "Reinitializing textures." << endl;
       InitTextures();
